+++ packages/services/src/fee_metrics_tracker.rs
@@ -1,3 +1,5 @@
 pub mod cache;
 pub mod port;
 pub mod service;
+#[cfg(feature = "test-helpers")]
+pub mod testing;
